1.测试企微每日运营，修复若干bug.

+++ operate/src/main/java/com/linksteady/operate/service/impl/QywxDailyServiceImpl.java
@@ -240,7 +240,7 @@ public class QywxDailyServiceImpl implements QywxDailyService {
         String taskDt = qywxDailyHeader.getTaskDateStr();
 
         //获取任务观察的天数
-        int effectDays = qywxDailyHeader.getEffectDays().intValue();
+        long effectDays = qywxDailyHeader.getEffectDays();
 
         //任务提交的日期
         LocalDate taskDtDate = LocalDate.parse(taskDt, DateTimeFormatter.ofPattern(dateFormat));
